backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
Pillow==8.3.1
djangorestframework>=3.12.4
djangorestframework-simplejwt
django-filter
django-rest-swagger